Moved read_graphfile function to io.py
